Allow validating patches for functions in same file
I noticed this problem while looking at #8. If you defined a function
and then defined a patch for it in the same file, then Emacs would
suddenly not be able to find the definition location for the function,
and el-patch-validate would therefore fail on it.

This was because the el-patch--stealthy-eval functioned by removing
entries from current-load-list. But if the function was defined in the
same file, the entry for the original definition of the function would
still be in current-load-list, and removing it would leave Emacs with
no recollection of how the function came to be defined! To fix the
issue, I made it so that entries are only removed from
current-load-list if they weren't there before the eval.