[#30] Correctly autoload el-patch
Previously, you could get an error:

    Unregistered definition type ‘defalias’

if you tried to evaluate an `el-patch-defun' form or similar inside of
an `eval-after-load' or similar, without having previously loaded
`el-patch'. This was due to an autoloading bug, which this commit
fixes.

In retrospect, the documentation I added when originally addressing
this problem was wrong -- it should not be necessary to explicitly
load a package in order to use its macros, at least in this
context (it is sometimes necessary in other cases, most notably when a
macro from package X is used inside an `eval-after-load' form for
package Y which is evaluated before package X is loaded, and the macro
from package X is autoloaded, and it does not evaluate all of its
arguments, and one of its arguments can be mistaken for an invocation
of some other macro from package Z).